requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cffi==1.17.1
Deprecated==1.2.18
discord.py==2.5.2
frozenlist==1.6.2
fuzzywuzzy==0.18.0
idna==3.10
lxml==5.4.0
multidict==6.4.4
packaging==25.0
pikepdf==9.8.1
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
python-dotenv==1.1.0
pytz==2025.2
RapidFuzz==3.13.0
typing_extensions==4.14.0
wrapt==1.17.2
xattr==1.1.4
yarl==1.20.0
